+++ benchmark/bin/write.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use
+
 import 'dart:async';
 
 import 'package:objectbox_benchmark/benchmark.dart';
@@ -10,11 +12,14 @@ void main() async {
   await Put().report();
   await PutInTx().report();
   await PutMany().report();
-  await PutAsyncSequential().report();
+  await PutQueuedAwaitResult().report();
+  await PutQueued().report();
+  await PutAsync().report();
+  await PutQueuedAwaitResultParallel().report();
+  await PutQueuedParallel().report();
   await PutAsyncParallel().report();
   await PutAsyncToList().report();
   await PutAsyncAwait().report();
-  await PutQueued().report();
   await RunInTx().report();
   await RunInTxAsync().report();
   await RunAsync().report();
@@ -50,22 +55,77 @@ class PutMany extends DbBenchmark {
 }
 
 /// Runs putAsync one-by-one, use to measure time of a single call.
-class PutAsyncSequential extends DbBenchmark {
+class PutQueuedAwaitResult extends DbBenchmark {
   static const count = 10;
   final items = prepareTestEntities(count, assignedIds: true);
 
-  PutAsyncSequential() : super('$PutAsyncSequential', iterations: count);
+  PutQueuedAwaitResult() : super('$PutQueuedAwaitResult', iterations: count);
+
+  @override
+  FutureOr<void> runIteration(int iteration) =>
+      box.putQueuedAwaitResult(items[iteration]);
+}
+
+/// Runs putQueued one-by-one, use to measure time of a single call.
+class PutQueued extends DbBenchmark {
+  static const count = 10;
+  final items = prepareTestEntities(count, assignedIds: true);
+
+  PutQueued() : super('$PutQueued', iterations: count);
+
+  @override
+  FutureOr<void> runIteration(int iteration) {
+    box.putQueued(items[iteration]);
+    store.awaitAsyncSubmitted();
+  }
+}
+
+/// Runs putAsync one-by-one, use to measure time of a single call.
+class PutAsync extends DbBenchmark {
+  static const count = 10;
+  final items = prepareTestEntities(count, assignedIds: true);
+
+  PutAsync() : super('$PutAsync', iterations: count);
 
   @override
   FutureOr<void> runIteration(int iteration) => box.putAsync(items[iteration]);
 }
 
+/// Runs many PutQueuedAwaitResult calls in parallel and waits until the last one completes,
+/// use to assert parallelization capability.
+class PutQueuedAwaitResultParallel extends DbBenchmark {
+  static const count = 100;
+  final items = prepareTestEntities(count, assignedIds: true);
+
+  PutQueuedAwaitResultParallel() : super('$PutQueuedAwaitResultParallel');
+
+  @override
+  Future<void> run() async =>
+      await Future.wait(items.map(box.putQueuedAwaitResult));
+}
+
+/// Runs many putQueuedParallel calls in parallel and waits until the last one
+/// completes, use to assert parallelization capability.
+class PutQueuedParallel extends DbBenchmark {
+  static const count = 100;
+  final items = prepareTestEntities(count, assignedIds: true);
+
+  PutQueuedParallel() : super('$PutQueuedParallel', iterations: 1);
+
+  @override
+  Future<void> run() async {
+    items.forEach(box.putQueued);
+    store.awaitAsyncSubmitted();
+  }
+}
+
 /// Runs many putAsync calls in parallel and waits until the last one completes,
 /// use to assert parallelization capability.
 class PutAsyncParallel extends DbBenchmark {
+  static const count = 100;
   final items = prepareTestEntities(count, assignedIds: true);
 
-  PutAsyncParallel() : super('$PutAsyncParallel');
+  PutAsyncParallel() : super('$PutAsyncParallel', iterations: 1);
 
   @override
   Future<void> run() async => await Future.wait(items.map(box.putAsync));
@@ -80,7 +140,7 @@ class PutAsyncToList extends DbBenchmark {
 
   @override
   Future<void> run() async {
-    final futures = items.map(box.putAsync).toList(growable: false);
+    final futures = items.map(box.putQueuedAwaitResult).toList(growable: false);
     await Future.wait(futures);
     store.awaitAsyncSubmitted();
   }
@@ -93,19 +153,7 @@ class PutAsyncAwait extends DbBenchmark {
 
   @override
   Future<void> run() async {
-    items.forEach((item) async => await box.putAsync(item));
-    store.awaitAsyncSubmitted();
-  }
-}
-
-class PutQueued extends DbBenchmark {
-  final items = prepareTestEntities(count, assignedIds: true);
-
-  PutQueued() : super('$PutQueued', iterations: count);
-
-  @override
-  Future<void> run() async {
-    items.forEach(box.putQueued);
+    items.forEach((item) async => await box.putQueuedAwaitResult(item));
     store.awaitAsyncSubmitted();
   }
 }