+++ objectbox/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## latest
 
-* Add `getAsync`, `getManyAsync` and `getAllAsync` to `Box`.
+* Added and updated async APIs in `Box`:
+  * new `getAsync`, `getManyAsync`, `getAllAsync`,
+  * new `putAsync` which supports objects with relations,
+  * renamed the former `putAsync` to `putQueuedAwaitResult`.
 * Add new async `Query` APIs: `findAsync`, `findFirstAsync`, `findUniqueAsync` and `findIdsAsync`.
 * Breaking changes to generated code: run `flutter pub run build_runner build` 
   (or `dart run build_runner build` for Dart Native apps) after updating!